선수CRUD기능 각 기능별로 나누기
기능별로 나누기
